+++ NEWS.md
@@ -1,4 +1,4 @@
-# rsconnect (development version)
+# rsconnect 1.2.0
 
 * Addressed a number of republishing and collaboration issues where the
   content was incorrectly published to a new location rather than reusing an